Base
Define start level 25 for heroes. Define respawn time of 10s. Save
original scale of heroes (Issue #1).
Put flags on map when start the game (Issue #3).
If a hero is holding a flag, then change scale of the hero and the level
of the flag every 20 seconds (Issue #6).
When a hero dies: remove flag from inventory of hero and respawn flag,
reset scale of hero (Issue #4).